# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REDPR документация АПИ",
        "description": "Документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/items/description": {
            "get": {
                "tags": [
                    "item"
                ],
                "summary": "Описание",
                "operationId": "description",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Категория",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Описание товара"
                    }
                }
            }
        },
...
        "/api/order/view": {
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "Посмотреть заказы",
                "description": "Посмотреть заказы",
                "operationId": "orderview",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order/view/{id}": {
            "post": {
                "tags": [
                    "order"
                ],
                "summary": "Посмотреть заказы",
                "description": "Посмотреть заказы",
                "operationId": "orderviewsingle",
                "requestBody": {
                    "description": "Апи Токен",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "api_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "6WxjM0XOruMPWPnJKEAPHNIMwNpe0bAU7iGWswoKrQDuXC5MNUmuJh1Y4GuG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CallBack с товаром",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}